Add script to rewrite test sets
